所有非 md 与 pages 文件，都定义为资源文件，原样复制到目标目录

+++ mkdocs_translator/utils.py
@@ -27,12 +27,18 @@ def copy_resources(source_dir: Path, target_dir: Path):
         source_dir: The source directory
         target_dir: The target directory
     """
-    non_translatable_patterns = ['*.png', '*.jpg', '*.jpeg', '*.gif', '*.css', '*.js']
+
+    # non_translatable_patterns = ['*.png', '*.jpg', '*.jpeg', '*.gif', '*.css', '*.js']
+    translatable_extensions = {'.md', '.pages'}
     
-    for pattern in non_translatable_patterns:
-        for source_file in source_dir.glob(f'**/{pattern}'):
+    for source_file in source_dir.glob('**/*'):
+        if source_file.is_dir():
+            continue
+            
+        if source_file.suffix not in translatable_extensions:
             relative_path = source_file.relative_to(source_dir)
             target_file = target_dir / relative_path
             
-            target_file.parent.mkdir(parents=True, exist_ok=True)
-            target_file.write_bytes(source_file.read_bytes()) 
+            if not target_file.exists():
+                target_file.parent.mkdir(parents=True, exist_ok=True)
+                target_file.write_bytes(source_file.read_bytes())
